feat: p230_2318【7.1.5-cpack专项】gradle仓库清理失效（maven仓库是生效的）

+++ src/backend/maven/biz-maven/src/main/kotlin/com/tencent/bkrepo/maven/event/EventConsumer.kt
@@ -3,9 +3,9 @@ package com.tencent.bkrepo.maven.event
 import com.tencent.bkrepo.common.artifact.event.base.ArtifactEvent
 import com.tencent.bkrepo.common.artifact.event.base.EventType
 import com.tencent.bkrepo.common.artifact.event.repo.RepositoryCleanEvent
-import com.tencent.bkrepo.maven.constants.REPO_TYPE
 import com.tencent.bkrepo.maven.service.MavenDeleteService
 import com.tencent.bkrepo.repository.constant.SYSTEM_USER
+import com.tencent.bkrepo.repository.pojo.packages.PackageType
 import org.springframework.stereotype.Component
 import java.util.function.Consumer
 
@@ -15,8 +15,11 @@ class EventConsumer (
 ) : Consumer<ArtifactEvent> {
     override fun accept(event: ArtifactEvent) {
         require(event.type == EventType.REPOSITORY_CLEAN) { return }
+        val packageType = event.data[RepositoryCleanEvent::packageType.name]
+        if (packageType != PackageType.MAVEN.name && packageType != PackageType.GRADLE.name) {
+            return
+        }
         with(event) {
-            require(data[RepositoryCleanEvent::packageType.name] == REPO_TYPE) { return }
             val versionList = data[RepositoryCleanEvent::versionList.name] as List<String>
             versionList.forEach {
                 service.deleteVersion(
